nitrogen6x: Migrate DDR_MB to Kconfig
Move the CONFIG_DDR_MB symbol to Kconfig.  A later clean-up would be to
make dynamic memory size detection work based on how this is done on
other i.MX6 platforms.

Cc: Troy Kisky <troy.kisky@boundarydevices.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
For example, mx6cuboxi does dynamic memory size detection and that code
could be adapted for this platform and perhaps moved to a more generic
imx6 location.