arm: Disable ATAGs support
With the exception of nokia_rx51 and icnova-a20-swac, disable ATAG
support.  A large number of platforms had enabled support but never
supported a kernel so old as to require it.  Further, some platforms are
old enough to support both, but are well supported by devicetree
booting, and have been for a number of years.  This is because some of
the ATAGs related functions have been re-used to provide the same kind
of information, but for devicetree or just generally to inform the user.
When needed still, rename these functions to get_board_revision()
instead, to avoid conflicts.  In other cases, these functions were
simply unused, so drop them.

Cc: Phil Sutter <phil@nwl.cc>
Cc: Stefan Bosch <stefan_b@posteo.net>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
There are some platforms, such as Synology that I'm worried should also
perhaps be keeping this enabled.  Also, after v1 of the series Stefan
sent a patch for friendlyarm nanopi2, but with this more board update
that particular path doesn't work.  I've updated things a little to keep
the spirit of that change.