doc: Add doc/develop/release_cycle.rst
Migrate the RelaseCycle wiki page to Sphinx.  In terms of visible
changes, we stop having a dynamic countdown to when the release is.  And
we drop the year-based statistics, that were not being kept up to date.
For the moment, we only link to statistics for v2022.07 but will add
back the historical data in a subsequent patch.

Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
Changes in v3:
- Minor rewords per Sean, on when we do point releases and about what
  the merge window means.
- Minor rewords per Heinrich, and try and make it clearer about why we
  have commented out rcN lines for the next release.
- Note that I have completed migration of all of the other stats pages,
  but as a single commit it is nearly 2MiB.

Changes in v2:
- New patch.  Note that I'm not including the v1.3.0 to v2022.04 stats
  here as while I can generate them, I want to make sure we're happy
  with how v2022.07 ones look as there is some manual effort involved.
  It's a reasonable amount of manual effort however for a single release.