CI: Azure: Merge PowerPC jobs in to one
At this point given the number of PowerPC platforms we have, a single
job to build them all fits within the time limit we have in Azure.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
This reduces the overall build time marginally due to freeing up some
parallel build slots. However, it needs the DM_ETH series to be applied
first as that removes a handful of PowerPC platforms and when we do them
all at once we're very close to the maximum run time limit.