rsa-verify: Rework host check for CONFIG_RSA_VERIFY_WITH_PKEY
While we do not want to use CONFIG_RSA_VERIFY_WITH_PKEY on the host, we
cannot undef the symbol in this manner. As this ends up only being a
single location in a file that already has other checks for HOST_CC, add
one more HOST_CC check instead.

Signed-off-by: Tom Rini <trini@konsulko.com>