kconfig: Add a configuration file for nospl
This file indicates which options are not for use in U-Boot SPL. It allows
the kconfig tool to correctly generate its output.

It was generated with

   ./tools/moveconfig.py --scan-source -u

Note that this only produces sensible results when used with the current
source code, as of this patch. Following patches adjust the code, e.g. to
drop use of CONFIG_IS_ENABLED(), so the program will not produce anything
useful when run on that version of the code.

Series-changes: 3
- Add DFU_VIRT FSL_ISBC_KEY_EXT GENERATE_SMBIOS_TABLE
  GMAC_ROCKCHIP I8259_PIC IMX_RDC NXP_FSPI PCI_PNP PHY_CADENCE_TORRENT
  QFW_PIO SEC_FIRMWARE_ARMV8_PSCI SPI_FLASH_BAR SPI_FLASH_MACRONIX
- Add CMD_BOOTEFI_BOOTMGR CMD_EFICONFIG CMD_PXE
- Fix move_config typo
- Add ARCH_VERSAL_NET CMD_DHCP ENV_WRITEABLE_LIST EXPO
- Add TARGET_KMCOGE5NE TARGET_KMETER1

Signed-off-by: Simon Glass <sjg@chromium.org>