kconfig: Adjust the meaning of CONFIG_IS_ENABLED()
Adjust this to require the CONFIG option be provided, so that instead of:

   CONFIG_IS_ENABLED(FOO)

you must write:

   CONFIG_IS_ENABLED(CONFIG_FOO)

This is in preparation for dropping this and just using it is as the new
implementation of IS_ENABLED(). For now, update IS_ENABLED() to use
CONFIG_IS_ENABLED(), so we can rely on the three-argument version.

Signed-off-by: Simon Glass <sjg@chromium.org>