Makefile: Drop SPL_ and SPL_TPL_ macros
These are not needed anymore with split configs. Drop them.

This also drops an incorrect $(SPL_TPL) in arch/powerpc/lib/Makefile

Series-changes: 3
- Move to conf_nospl: CMD_DHCP DFU_VIRT GENERATE_SMBIOS_TABLE
- Move to conf_nospl: GMAC_ROCKCHIP I8259_PIC IMX_RDC NXP_FSPI PCI_PNP
- Move to conf_nospl: PHY_CADENCE_TORRENT QFW_PIO SEC_FIRMWARE_ARMV8_PSCI
- Move to conf_nospl: SPI_FLASH_BAR SPI_FLASH_MACRONIX
- Move to conf_nospl: SCMD_BOOTEFI_BOOTMGR CMD_EFICONFIG CMD_PXE

Series-changes: 2
- Reworked based on earlier patches

Series-changes: 2
- Move patches from splb which add new Kconfig options

Series-to: u-boot
Series-version: 3
Series-cc: trini
Cover-letter:
RFC: Migrate to split config
U-Boot uses an SPL prefix on CONFIG options to indicate when an option
relates to SPL. For example, while CONFIG_TEXT_BASE is the text base for
U-Boot proper, CONFIG_SPL_TEXT_BASE is the text base for SPL.

Within the code it is possible do things like CONFIG_VAL(TEXT_BASE) to
get that value. It returns the appropriate option, depending on the phase
being built.

The same applies for boolean options, like CONFIG_BLK and CONFIG_SPL_BLK
but in that case we use CONFIG_IS_ENABLED(BLK) to obtain the value.

For Makefiles we use an SPL_TPL_ macro which evalutes either to empty,
SPL, TPL or even VPL depending on the build phase. So we can do things
like:

   obj-$(CONFIG_$(SPL_TPL_)CLK) += clk-uclass.o

To complicate things a little, some options like CONFIG_ARCH_APPLE do not
have an SPL version, so people have to remember to use IS_ENABLED()
instead of CONFIG_IS_ENABLED(). If the latter is used, the value will be
false.

All of this is rather confusing, since:

- people have to know whether to use IS_ENABLED() or CONFIG_IS_ENABLED()
- the Makefiles are cluttered with $(SPL_TPL_) constructs
- plain use of a CONFIG option (e.g. #ifdef CONFIG_CLK) may be correct in
  some cases but not in others

This series updates U-Boot to use separate config files for each phase of
the build. It drops use of CONFIG_IS_ENABLED() and SPL_TPL_ as well.

For now this series does not fully work due to inconsistencies in the
Kconfig options. Many of these have been fixed [1] [2a] [2b] but more
remain.

This series is available at u-boot-dm/splc-working

Size changes are available here: https://pastebin.com/hSjhM3y7

[1] https://patchwork.ozlabs.org/project/uboot/list/?series=339004
[2a] https://patchwork.ozlabs.org/project/uboot/list/?series=339199&state=*
[2b] https://patchwork.ozlabs.org/project/uboot/list/?series=339208&state=*
END

Signed-off-by: Simon Glass <sjg@chromium.org>