imx: watchdog: Correct watchdog error condition
This is intended to make sure that one of the conditions is met, but
causes a build error with imx6qdl_icore_mmc.

Fix it by using the PPL condition instead, so that nothing changes when
building for SPL with split config.

In fact we should drop the old watchdog code and just use the new
watchdog uclass.

Signed-off-by: Simon Glass <sjg@chromium.org>