freescale: Fix odd use of ESDHCI_QUIRK_BROKEN_TIMEOUT_VALUE
This is not a CONFIG option so we should not be using IS_ENABLED() on it,
particularly not when it is not defined to anything, so shows up as
calling IS_ENABLED() with no arguments.

Just check it normally.

This fixes a build error with split config on T2080QDS.

Signed-off-by: Simon Glass <sjg@chromium.org>