Allow distroboot environment to be the same in SPL
With the split config, we find that CONFIG_CMD_DHCP is not enabled in SPL.
While this is normally correct, it causes a build error with
BOOT_TARGET_DEVICES_references_PXE_without_CONFIG_CMD_DHCP_or_PXE

This is because some boards include PXE and DHXP in their
BOOT_TARGET_DEVICES #define. Before split config CONFIG_CMD_DHCP is
enabled in SPL, but with split config it is not, since we don't support
any commands in SPL.

The easiest way to fix this seems to be to use the CONFIG_PPL_... form.
This retains the existing behaviour.

Series-changes: 3
- Add new patch to allow distroboot environment to be the same in SPL