mx9: Correct repeatable build error
For some reason every second time imx93_11x11_evk is built it gives an
error:

   make O=/tmp/x BINMAN_ALLOW_MISSING=1

It seems to sometimes skip generation of the .cfgout file and then
eventually Binman complains:

   ValueError: Error 1 running 'mkimage -d ./mkimage.spl.mkimage -n
      spl/u-boot-spl.cfgout -T imx8image -e 0x2049A000
      ./mkimage-out.spl.mkimage': Fail open first container file
      mx93a1-ahab-container.img

Correct this by using if_changed instead of if_changed_dep

The only reason this hasn't come up in CI is that buildman did not retry
failing builds of current source, but now it does.

Note: The logic in this Makefile should be moved to Binman, e.g. these
warnings duplicate Binman functionality:

  WARNING 'bl31.bin' not found, resulting binary may be not-functional
  WARNING 'tee.bin' not found, resulting binary may be not-functional

Signed-off-by: Simon Glass <sjg@chromium.org>