Revert "buildman: Always use the full path in CROSS_COMPILE"
There are operations in buildman that result in running the cross-tools
(such as performing size checks) and now that we have not modified PATH
to know where our tools are, these operations fail.

This reverts commit 6c0a3cf75f72370deec3ee516a9dd377397af207.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
I'm likely to apply this revert fairly quickly as it turns out (and I
assumed incorrectly it was just due to locality impact, but I started
getting worried this morning, sigh) none of my size check builds have
been really reporting true values.

Cc: Simon Glass <sjg@chromium.org>