zlib: Fix inflate_fast() when POSTINC macro is set
inflate_fast() implements possibility to be build with
POSTINC set or unset. That POSTINC macro was introduced by
commit e89516f031db ("zlib: split up to match original source tree")
but forced to #undef in zlib.h.

However, in the meantime that commit removed UP_UNALIGNED() macro
introduced by commit cd514aeb996e ("zlib: Optimize decompression")
that was meant to implement proper behaviour depending on POSTINC.

In addition, commit cd514aeb996e ("zlib: Optimize decompression") has
a bug for when POSTINC is set in the following pointer arithmetic:

	pat16 = *(sout-2+2*OFF)

The intension is to remove 2 bytes when OFF is 0, but sout being a
short, this leads to removing 4 bytes instead of 2. The bug in not
visible when OFF is set to 1 which has always been the case until now.

In preparation for followup patch that will activate POSTINC, add back
UP_UNALIGNED() macro and fix the above pointer calculation.

Fixes: cd514aeb996e ("zlib: Optimize decompression")
Cc: Michal Simek <michal.simek@amd.com>
Signed-off-by: Christophe Leroy <christophe.leroy@csgroup.eu>