turris_1x: Normalize Kconfig usage
We don't set many options in the board Kconfig entry file but instead
use defconfigs, select in some cases on the target itself, or update the
"default" options of the main entries when needed. In this case we can
remove most of the board Kconfig entries and just add them to the
defconfig like other platforms.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
I'd be open to someone else taking this patch and doing a v2 that has
one of the defconfig's #include the other and then make the required
changes (see other defconfigs which use #include for the constraints
however) as an alternative.

Cc: "Marek Mojík" <marek.mojik@nic.cz>
Cc: "Marek Behún" <kabel@kernel.org>