configs: replace '# CONFIG_NET is not set' with CONFIG_NO_NET=y
The parent patch has made the networking stack a choice between NO_NET,
NET and NET_LWIP. Therefore '# CONFIG_NET is not set' is now
'CONFIG_NO_NET=y'. Adjust the defconfigs accordingly. Note that this
patch is intended to be folded in but is kept separate separate for now
to make review easier.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Felix Brack <fb@ltec.ch>