bloblist: Introduce BLOBLIST_PRIOR_STAGE options
Introduce an option to control if we expect that a prior stage has
created a bloblist already and thus it is safe to scan the address. We
need to have this be configurable because we do not (and cannot) know if
the address at CONFIG_BLOBLIST_ADDR is in regular DRAM or some special
on-chip memory and so it may or may not be safe to read from this
address this early.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Simon Glass <sjg@chromium.org>