Gitlab: Add missing symlink for qemu_arm64_lwip boardenv file
When adding the symlink for the conf file so qemu_arm64_lwip uses
qemu_arm64 configuration information, the symlink for the boardenv file
was missed in Gitlab (but not Azure). Add that in now.

Fixes: fd10d156db3f ("CI: add qemu_arm64_lwip to the test matrix")
Signed-off-by: Tom Rini <trini@konsulko.com>
---
I'll take this through my tree once reviewed, no need for a PR for just
this.

Cc: Jerome Forissier <jerome.forissier@linaro.org>