doc: Include pytests directly in documentation
As there is support within the Sphinx framework to take Python
docstrings and generate documentation from it, we can make use of that
to more clearly and consistently document our pytests and how to
configure them. To do this we:
- Add PYTHONPATH to DOC_TARGETS so that modules will be found
- Add the autodoc extension for Python and then redirects so we can move
  py_testing.html (which is heavily referenced in archives, etc).
- Create a doc/develop/pytest directory and move the existing py_testing
  file there, update references.
- Update the test_000_version.py test to have a docstring
- Move the imports in test_net_boot.py so the docstring is in the
  correct place.
- Use code-block for all of the code blocks in test_net_boot.py so they
  are rendered correctly in documentation.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cc: Simon Glass <sjg@chromium.org>

As this is a RFC, I didn't split it up as much as it needs to be as I'm
looking for some feedback on the concept.