arc: Rework Kconfig choice menus
Currently we get a number of warnings about using "select" statements
on "choice" options and that this as no effect. Rework some of these so
that we use "depends on" rather than select for the CPU, and have the
CPU select the MMU version. The MMU version is no longer prompted for.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Alexey Brodkin <alexey.brodkin@synopsys.com>
Cc: Eugeniy Paltsev <Eugeniy.Paltsev@synopsys.com>
Cc: uboot-snps-arc@synopsys.com