atmel: Remove local <dt-bindings/pinctrl/at91.h>
We have this binding file in both include/dt-bindings/ and
dts/upstream/include/dt-bindings. These files are identical save for the
fact that we do not have commit adb2424d0d05 ("dt-bindings: clock: add
clock definitions for Ralink SoCs") from the kernel applied. However,
this change is rather important and should not have been omitted here
most likely. Remove our local copy to get in sync with upstream now.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Eugen Hristev <eugen.hristev@linaro.org>
Cc: Mihai Sain <mihai.sain@microchip.com>

The full kernel commit log is at
https://web.git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=0b3292852863215825f88905b9dbafc3101e1d7e
and so this sounds very important. Please let me know if this should go
in master for v2025.07 (or Eugen, please pick it and send a PR). Thanks!