scripts/Makefile.lib: Rework upstream_dtsi_include to use find
A problem with the logic in upstream_dtsi_include currently is that it
does not list directories such as dts/upstream/src/arm/nxp/imx and so
will not findi "imx6ul-pinfunc.h" for example as it is normally and
correctly included without vendor sub-paths.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
This becomes a problem if we remove identical header files from
arch/arm/dts for example and only always use the upstream one.

Cc: Sumit Garg <sumit.garg@kernel.org>