timer: Tighten CADENCE_TTC_TIMER dependencies
This driver implements timer_get_boot_us() which is global namespace and
can conflict with other implementations of the function for example in
"allyesconfig". Start by limiting this driver to platforms which use it.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
I know that the help says Zynq or ZynqMP but only the ZYNQMP_R5
architecture enables this today. I'm happy to expand this dependency
more however or let someone from AMD v2 this.

Cc: Michal Simek <michal.simek@amd.com>